関数をutil/string_exに移動
変更対象の関数を、ビルドできるファイルに移動し、コードを直接テストできるようにする
