+++ sakura_core/util/string_ex.h
@@ -337,5 +337,9 @@ inline int wcsncmp_auto(const wchar_t* strData1, const wchar_t* szData2)
 	#define _tcsncmp_literal strncmp_literal
 #endif
 
+BOOL IsURL( const wchar_t*, int, int* );/* 指定アドレスがURLの先頭ならばTRUEとその長さを返す */
+BOOL IsMailAddress( const wchar_t*, int, int* );	/* 現在位置がメールアドレスならば、NULL以外と、その長さを返す */
+
+
 #endif /* SAKURA_STRING_EX_29EB1DD7_7259_4D6C_A651_B9174E5C3D3C9_H_ */
 /*[EOF]*/
