+++ sakura_core/view/colors/CColor_Url.cpp
@@ -5,6 +5,7 @@
 #include "doc/CEditDoc.h"
 #include "doc/layout/CLayout.h"
 #include "types/CTypeSupport.h"
+#include "util/string_ex.h" // IsMailAddress
 
 
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
